src/requirements.txt
ultralytics
opencv-python-headless
matplotlib
requests
python-dotenv
cvzone
boto3